# spec/api/openapi.yaml
openapi: 3.1.0
info:
  title: RGPerp API
  version: 0.1.0
  description: |
    链上托管、链下交易、链下风控、链下清算、外部对冲永续合约系统 API 规范。
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string
        data: {}
      required: [code, message, trace_id]
...
    NonceRequest:
      type: object
      properties:
        address:
          type: string
        chain_id:
          type: integer
      required: [address, chain_id]
    NonceResponse:
      type: object
      properties:
        nonce:
          type: string
        domain:
          type: string
        chain_id:
          type: integer
        expires_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      properties:
        address:
          type: string
        chain_id:
          type: integer
        nonce:
          type: string
        signature:
          type: string
      required: [address, chain_id, nonce, signature]
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        evm_address:
          type: string
        status:
          type: string
      required: [id, evm_address, status]
...
paths:
  /api/v1/auth/nonce:
    post:
      tags: [Auth]
      summary: 获取登录 nonce
      parameters:
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NonceRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NonceResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 钱包签名登录
      parameters:
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
...